LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.3_Future_Reserved_Words/S7.5.3_A1.16-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected use of reserved word 'import'
S7.5.3_A1.16

PASS Expected parsing failure

TEST COMPLETE
